orchestrators/user-orchestrator/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
httpx==0.27.0
tenacity==8.5.0
python-dotenv==1.0.1
prometheus-fastapi-instrumentator==7.1.0
